eventsAPI/requirements.txt
aniso8601==6.0.0
attrs==19.1.0
certifi==2019.3.9
chardet==3.0.4
Click==7.0
Flask==1.0.3
Flask-Cors==3.0.8
flask-restplus==0.12.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
jsonschema==3.0.1
MarkupSafe==1.1.1
meetup-api==0.1.1
pyrsistent==0.15.2
pytz==2019.1
PyYAML==5.1.1
requests==2.22.0
six==1.12.0
urllib3==1.25.3
Werkzeug==0.15.4
flask_basicauth==0.2.0